+++ fdr-impl/src/main/java/com/dwarfeng/fdr/impl/handler/bridge/PersisterOnlyBridge.java
@@ -3,6 +3,7 @@ package com.dwarfeng.fdr.impl.handler.bridge;
 import com.dwarfeng.fdr.stack.bean.dto.FilteredData;
 import com.dwarfeng.fdr.stack.bean.dto.NormalData;
 import com.dwarfeng.fdr.stack.bean.dto.TriggeredData;
+import com.dwarfeng.fdr.stack.exception.KeeperNotSupportedException;
 import com.dwarfeng.subgrade.stack.exception.HandlerException;
 
 /**
@@ -21,40 +22,18 @@ public abstract class PersisterOnlyBridge extends AbstractBridge {
     }
 
     @Override
-    public boolean supportKeeper() {
-        return false;
-    }
-
-    @Override
     public Keeper<NormalData> getNormalDataKeeper() throws HandlerException {
-        try {
-            throw new IllegalStateException("不支持的操作");
-        } catch (Exception e) {
-            throw new HandlerException(e);
-        }
+        throw new KeeperNotSupportedException();
     }
 
     @Override
     public Keeper<FilteredData> getFilteredDataKeeper() throws HandlerException {
-        try {
-            throw new IllegalStateException("不支持的操作");
-        } catch (Exception e) {
-            throw new HandlerException(e);
-        }
+        throw new KeeperNotSupportedException();
     }
 
     @Override
     public Keeper<TriggeredData> getTriggeredDataKeeper() throws HandlerException {
-        try {
-            throw new IllegalStateException("不支持的操作");
-        } catch (Exception e) {
-            throw new HandlerException(e);
-        }
-    }
-
-    @Override
-    public boolean supportPersister() {
-        return true;
+        throw new KeeperNotSupportedException();
     }
 
     @Override